check if file exists or not, if already exist, dont download

+++ download_utils.py
@@ -66,7 +66,9 @@ def down_fr_url(urls: list, save_dir: str='', unzip: bool=False):
         try:
             fn = retri_fn_url(url)
             save_path = os.path.join(save_dir, fn)
-            
+            if os.path.exists(save_path) and os.path.getsize(save_path)>=retri_file_size(url):
+                print('{} already exists.'.format(save_path))
+                continue
             print('Downloading {} ...'.format(fn))
             urlretrieve(url, save_path)
             print('\n')
